+++ Wire-iOS/Sources/UserInterface/ContactsUI/ContactsViewController+ShareContactsViewControllerDelegate.swift
@@ -21,7 +21,7 @@ import Foundation
 extension ContactsViewController: ShareContactsViewControllerDelegate {
     func shareDidFinish(_ viewController: UIViewController) {
         // Reload data source
-        dataSource.search(withQuery: "", searchDirectory: dataSource.searchDirectory)
+        dataSource.searchQuery = ""
         dismissChildViewController(viewController)
     }
